Fixed a security bug in the three party protocol
Previously there was nothing signed by the meter which tied the measurement to a particular week and so a cheating customer could replay low measurements.

To fix this, the time information signed by the smart meter is no longer limited to be an hour of the week. Instead this should be hours since the meter first began measuring. In a real system the provider needs to check that it has all of the measurements it expects and that they are for the correct hours. This is **NOT** implemented here as it would be a pain for the interactive demo.